Refactor components into directories by concern
