+++ react-app/src/components/Header/Nav/SearchButton.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-import { ReactComponent as SearchIcon } from "./assets/search-solid.svg";
+import { ReactComponent as SearchIcon } from "../../assets/search-solid.svg";
 
 const SearchButtonStyled = styled.div`
   /* align-self: center; */
