+++ react-app/src/components/Header/Nav/index.js
@@ -7,7 +7,7 @@ import styled from "styled-components";
 import SearchButton from "./SearchButton";
 import UserPanel from "./UserPanel";
 import LanguagePicker from "./LanguagePicker";
-import { ReactComponent as HamburgerIcon } from "./assets/bars-solid.svg";
+import { ReactComponent as HamburgerIcon } from "../../assets/bars-solid.svg";
 
 const NavStyled = styled.nav`
   background-color: white;
